requirements.txt
alpaca-py
numpy
pandas
plotly